Refactor review utilities into package
