+++ tests/test_blank_code_reload.py
@@ -1,6 +1,6 @@
 import pandas as pd
 from decimal import Decimal
-from wsm.ui.review_links import _save_and_close
+from wsm.ui.review.io import _save_and_close
 from wsm.utils import _clean
 
 class DummyRoot:
